[PFS-25] Remove path from putFileURLTask proto.
