+++ src/server/pfs/server/pfsserver.proto
@@ -53,7 +53,6 @@ message PutFileURLTask {
   string dst = 1;
   string datum = 2;
   string URL = 3;
-  repeated string paths = 4;
   string start_path = 5;
   string end_path = 6;
   int64 start_offset = 7;
